Refactored listener classes to separate modules for cleaner implementation
